pyproject.toml
[project]
name = "autodev"
version = "0.1.0"
description = "Autonomous Development System based on Anthropic's long-running agent harness"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AutoDev Team"}
]
keywords = ["ai", "agent", "automation", "development"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "click>=8.0",
    "jinja2>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "playwright>=1.40.0",
]
e2e = [
    "playwright>=1.40.0",
]

[project.scripts]
autodev = "cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "config", "templates"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "e2e: end-to-end tests (deselect with '-m \"not e2e\"')",
    "integration: integration tests",
]
